+++ examples/next13-typescript-simple/app/client-config.ts
@@ -1,5 +1,3 @@
-
-
 export const clientConfig = {
   apiKey: process.env.NEXT_PUBLIC_FIREBASE_API_KEY!,
   authDomain: process.env.NEXT_PUBLIC_FIREBASE_AUTH_DOMAIN!,